add(application): 원서 수정 로직 트랜잭션 분리 및 보상 로직 추가
