+++ src/main/java/team/themoment/hellogsmv3/domain/application/entity/abs/AbstractMiddleSchoolGrade.java
@@ -2,10 +2,7 @@ package team.themoment.hellogsmv3.domain.application.entity.abs;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
+import lombok.*;
 
 import java.math.BigDecimal;
 import java.util.UUID;
@@ -14,6 +11,7 @@ import java.util.UUID;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Inheritance(strategy = InheritanceType.SINGLE_TABLE)
+@ToString
 public abstract class AbstractMiddleSchoolGrade {
 
     @Id
@@ -27,5 +25,9 @@ public abstract class AbstractMiddleSchoolGrade {
     public AbstractMiddleSchoolGrade(UUID id, @NonNull BigDecimal percentileRank) {
         this.id = id;
         this.percentileRank = percentileRank;
+    }
+
+    public void initId() {
+        this.id = null;
     }
 }
