+++ src/main/resources/application.yml
@@ -61,3 +61,9 @@ auth:
     - '*'
   loginEndPointBaseUri: /auth/v3/oauth2/authorization
   loginProcessingUri: /login/oauth2/code/*
+
+
+logging:
+  level:
+    org.hibernate.sql: debug
+    org.hibernate.type: trace
